// client/src/apiClient/Api.ts
export interface CreateClientDto {
  name: string;

  /** @format date-time */
  dob: string;
}